\name{crc32cExample-package}
\alias{crc32cExample-package}
\alias{crc32cExample}
\docType{package}
\title{\packageTitle{crc32cExample}}
\description{\packageDescription{crc32cExample}}
\section{Package Content}{\packageIndices{crc32cExample}}
\author{\packageAuthor{crc32cExample}}
\section{Maintainer}{\packageMaintainer{crc32cExample}}
\keyword{package}
