% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crc32cdemo}
\alias{crc32cdemo}
\title{Demo of calling 'crc32c' functions}
\usage{
crc32cdemo()
}
\value{
The function is called for its side effect of printing to stdout,
nothing is returned
}
\description{
Demo of calling 'crc32c' functions
}
